requirements.txt
#django-cms>=3.5,<3.6
-e git+git@github.com:divio/django-cms.git#egg=django-cms

# django CMS plugins and addons

djangocms-text-ckeditor>=3.6.0
djangocms-link>=2.1
djangocms-style>=2.0
djangocms-googlemap>=1.1
djangocms-snippet>=2.0
djangocms-video>=2.0
djangocms-file>=2.0,<3.0
djangocms-picture>=2.0,<3.0
django-filer>=1.3
djangocms-bootstrap4==1.0.0
djangocms-admin-style>=1.2,<1.3

# Django dependencies

Django<2.0
django-classy-tags>=0.7
django-sekizai>=0.9
django-treebeard>=4.0,<5.0

# Python dependencies

html5lib>=0.999999,<0.99999999
Pillow>=3.0
pytz
six
easy_thumbnails

# polls

-e git+git@github.com:divio/django-polls.git#egg=django-polls